backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
scikit-learn
xgboost
prophet
sqlalchemy
redis
psycopg2-binary
